Initial Commit
Add all the files to repository. This was made for my Databse class in my group of myself and 2 others. I had written all of the PHP and most of the HTML/CSS. My two teammembers (Emilio Sillart and Rimsha Rashid) had made the database and input all of the information. Emilio added the Google Map to ArtifactDirectory.html and Rimsha inserted the list of artifacts on ArtifactDirectory.html.